requirements.txt
# Dependências para a aula de agentes
crewai>=0.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20
python-dotenv>=1.0.0
google-search-results>=2.4.2
